Abort names merging on conflicts
Extractor considers all the text provided to be tightly coupled and merges all the names inside text (because considers that same names are referred to the same person).
Names can be more or less general. For example 'Steve Jobs' is more specific than 'Steve'.
When analyzing the text extractor will merge 'Steve' and 'Steve Jobs' to the same name ('Steve Jobs') that just meets two times.
There are some special case however. If 'Steve Jobs' and 'Steve Ballmer' meets in the same text, that will be not merged, obviously. But if just 'Steve' will meet in the text along with  'Steve Jobs' and 'Steve Ballmer' it is not clear what 'Steve' it refers to. So extractor will generate separate `Just 'Steve'` instance and provide some diagnostics